refactor: Separate LeadingNodes class
